Factor tt::has_inequivalence to its own file
